### 2018-02-25 #594 @whedon
[PRE REVIEW]: DynamicalSystems.jl: A Julia software library for chaos and nonlinear dynamics
**Submitting author:** @Datseris (<a href="http://orcid.org/0000-0002-6427-2385">George Datserus</a>)
**Repository:** <a href="https://github.com/JuliaDynamics/DynamicalSystems.jl" target ="_blank">https://github.com/JuliaDynamics/DynamicalSystems.jl</a>
**Version:** v0.11.0
