### 2018-02-25 #594 @Datseris
Both packages are part of the same GitHub organization, "JuliaDynamics" and I am the author of both of them.

As far as what they do, they are not related though. One is about 2-dimensional billiards, a quite special class of chaotic systems, while the one submitted here is about studying nonlinear dynamics and chaos.

Both packages certainly fall under the same field of science (Nonlinear Dynamics), but their underlying scope *and* code are very different.
