+++ docs/src/chaos/nlts.md
@@ -184,6 +184,6 @@ tight_layout()
 ```
 ![Broomhead-King example](https://i.imgur.com/xVWDjuh.png)
 
-we have used the same system as in the [delay coordinates reconstruction](definition/reconstruction) example, and picked the optimal
+we have used the same system as in the [delay coordinates reconstruction](/definition/reconstruction) example, and picked the optimal
 delay time of `τ = 30`. Regardless, the vanilla delay coordinates fail spectacularly
 when compared with the Broomhead-King coordinates.
