### 2018-02-25 #594 @ahwillia
I'm happy to review another submission! I'm busy for the next 2.5 weeks but
I can do a review after that.

If you want a quicker turn around I'd suggest finding someone else...

On Feb 25, 2018 11:02 AM, "George Datseris" <notifications@github.com>
wrote:

> @ahwillia <https://github.com/ahwillia> @mlxd <https://github.com/mlxd>
> have reviewed a previous submission of mine.
>
> Both the reviewers and myself were pleased with how smoothly the review
> process went, so I am suggesting them for this submission as well.
>
