### 2018-02-25 #594 @Datseris
Well, if I can have any say, then of course I'd prefer the faster start on the review :)
