### 2018-02-27 #598 @kyleniemeyer
@Datseris @dhhagan @ahwillia @mlxd 👋 the review will take place here.

@dhhagan you are the "main" reviewer, so please fill out the checklist above, but @ahwillia @mlxd please also see if you agree that those points were satisfied (or not) and comment appropriately.

Thanks!
