+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,13 @@
+Please follow the below guidelines, depending on whether you are
+reporting an issue or you are requesting a new feature.
+
+# Issue description OR description of requested feature
+mplah mplah.
+
+# Expected behavior OR description of possible implementation
+mplah.
+
+# MWE (only for issue reports)
+MWE stands for Minimal Working Example.
+
+Please include a piece of code that can replicate your problem in a simple example. This will help us identify the problem much faster.
