+++ CONTRIBUTING0.md
@@ -1,0 +1,1 @@
+See https://juliadynamics.github.io/DynamicalSystems.jl/latest/contributors_guide/.
