### 2018-03-03 e52a217
New docs version 0.12
