### 2018-03-05 JuliaDynamics/DynamicalSystems.jl#62 @dhhagan
Missing DOI for Paper in paper.bib
Please follow the below guidelines, depending on whether you are
reporting an issue or you are requesting a new feature.

# Issue description OR description of requested feature
There is a missing DOI for one of the papers listed in your bibliography; I believe the appropriate one can be found [here](http://adsabs.harvard.edu/abs/1988JPhA...21L.127K)
