### 2018-03-05 #598 @dhhagan
Hi @Datseris; everything looks great with the exception of one missing DOI as mentioned in [Issue 62](https://github.com/JuliaDynamics/DynamicalSystems.jl/issues/62). I believe this is the last minor hold-up!
