### 2018-03-05 8a62fc9
add doi for Kantz
Closes #62
