+++ paper/paper.bib
@@ -38,6 +38,7 @@ year = {1963}
   volume={21},
   number={3},
   pages={L127},
+  doi={10.1088/0305-4470/21/3/003},
   url={http://stacks.iop.org/0305-4470/21/i=3/a=003},
   year={1988},
   abstract={The authors investigate numerically how typical trajectories fill the phase space in low-dimensional symplectic (Hamiltonian) maps with finite phase space. They do not find any sign of a 'chaos threshold' as reported by other authors when the non-linearity parameters are increased. Instead, as expected from Arnold diffusion, they find that single trajectories fill most (if not all) of the coarse-grained phase space even for very small non-linearities. Due to the 'stickiness' of tori also observed in two-dimensional maps, this filling is much slower than what one might expect naively and is possibly described by power laws. The 'chaos threshold' observed in a previous paper is explained as a trivial effect.}
