### 2018-03-05 #598 @Datseris
Hello @dhhagan. Thank you for the kind words. I fixed your reported issue with the latest commit.

@kyleniemeyer I was wondering, whether it is possible to associate the paper with version v0.12 of the software, instead of the current one v0.11.

During the time this review was underway, a new feature was added (which is significant enough for a new version). I am asking you now, while v0.12 is still not tagged.
