### 2018-03-05 3154057
restructure introduction
fixes #63
