+++ docs/mkdocs.yml
@@ -1,6 +1,6 @@
 site_name:           DynamicalSystems.jl
 repo_url:            https://github.com/JuliaDynamics/DynamicalSystems.jl
-site_description:    A Julia package for the exploration of chaos in continuous and discrete dynamical systems.
+site_description:    A Julia software library for the exploration of chaos and nonlinear dynamics.
 site_author:         George Datseris
 
 theme:
