### 2018-03-07 690e032
note for differentiability of f
