### 2018-03-07 7fffe94
fix dollar signs
