+++ docs/src/definition/general.md
@@ -1,12 +1,12 @@
 # Dynamical Systems
 Currently a system in **DynamicalSystems.jl** can be either continuous
-$$
+```math
 \dot{\vec{u}} = \vec{f}(\vec{u}, t),
-$$
+```
 or discrete
-$$
+```math
 \vec{u}_{n+1} = \vec{f}(\vec{u}_n, t).
-$$
+```
 
 Keep in mind that almost all functions of **DynamicalSystems.jl** assume that $f$ is differentiable!
 
