+++ docs/src/tsprediction/localmodels.md
@@ -1,0 +1,27 @@
+# Predicting TImeseries or whatever
+mplah mplah
+this is a general thing about predictin timeseries
+
+## Local Model Prediction
+```@docs
+predict_timeseries
+AbstractLocalModel
+```
+### Example
+
+TODO FIX THIS SO THAT IT ACTUYALLY WORKS
+
+PUT THE PLOT WITH THE THREE CURVES HERE.
+
+(AND CODE)
+
+## Error Measures
+```@docs
+MSE1
+MSEp
+```
+
+### Estimating Parameters
+
+put here the code from `estimate_param`.
+and also a working example of course.
