+++ docs/mkdocs.yml
@@ -50,6 +50,8 @@ pages:
   - Periodicity: chaos/periodicity.md
   - Chaos Detection: chaos/chaos_detection.md
   - Visualization: chaos/visualization.md
+- Timeseries Prediction:
+  - Local Prediction: tsprediction/localmodels.md
 - Advanced Documentation: advanced.md
 - Contributor Guide: contributors_guide.md
 - News: news.md
