### 2018-03-12 #598 @kyleniemeyer
@Datseris that's no problem, we expect the version to change after reviewer comments.

However, if it's a significant change, then I would ask that @dhhagan take another look, with that new feature in mind.
