### 2018-03-12 #598 @Datseris
There is no reason for a second look, the feature was already in the
documentation before the reviewer started the review (because it comes from
another smaller package which is a dependency )
