### 2018-03-12 #598 @ahwillia
I have it on my calendar to take a look at it this on Thursday. What would be most useful for me to focus on?

@Datseris - can you provide a short discussion about how this library differs from [DifferentialEquations.jl](https://github.com/JuliaDiffEq/DifferentialEquations.jl)?

@ChrisRackauckas would you also mind weighing in on this briefly? Thanks so much.
