### 2018-03-12 #598 @kyleniemeyer
@ahwillia great, thanks! If you have time, I'd prefer for you to give the whole thing (software + paper) a look-over, independent of @dhhagan's review. At minimum, your thoughts on the software and whether it does what it advertises (etc.) would be helpful.
