### 2018-03-12 #598 @Datseris
@ahwillia I am not sure to respond because I don't understand the question.

Can you explain what you are asking about?

(your question implies that I have to somehow justify some difference
between the submitted package and differential equations, but that would
not make sense )
