### 2018-03-12 #598 @ChrisRackauckas
@ahwillia there's not really a similarity. DifferentialEquations.jl and JuliaDiffEq develops solvers and modeling packages. DynamicalSystems.jl has tooling for analysis of dynamical systems, such as determining Lyopunov exponents. It internally uses DifferentialEquations.jl to do the evolution of the systems, but that's only a small part of what's required for its algorithms.
