### 2018-03-12 #598 @ahwillia
Thanks Chris, this was my basic understanding. I have not had the chance to dig too deeply into this yet, but I could see new users being a bit confused.

@Datseris -- thank you for the detailed section on "Similar Existing Software". Would you mind adding a paragraph there or somewhere in the README / Documentation to make it clear to users that the numerical solvers are provided by `DifferentialEquations.jl` and the purpose of this package is to help characterize the behavior of solutions produced from said solvers? Something as simple as:

> While many libraries exist for numerically solving differential equations (CITATION DUMP), only a few libraries focus on X, Y, and Z.

The problem is not to "justify some difference" but to make the purpose of the library as clear as possible.
