### 2018-03-12 #598 @Datseris
It is already clear to the users that we are using DifferentialEquations.jl : https://juliadynamics.github.io/DynamicalSystems.jl/latest/definition/general/#DynamicalSystemsBase.DynamicalSystem

The documentation string of the struct `DynamicalSystem` is absolutely crucial for the entire library and I assume that any user that wants to use the non-data aspect of DynamicalSystems.jl has to have read it at least once. Thus, the usage of DifferentialEquations.jl is fully transparent.

*(not only that, but it is also the first documentation string of the library, so it is pretty hard to miss)*

> and the purpose of this package is to help characterize the behavior of solutions produced from said solvers?

This is very far way from the truth. I would say that only 1/10th of the total exported functions actually use the "solutions produced from said solvers". The rest 9/10th either does pure data analysis using Base Julia (and StaticArrays), or uses NearestNeighbors, or other packages, or internal solvers for discrete systems. Don't get me wrong, DifferentialEquations.jl is spectacular, but saying that all of DynamicalSystems.jl simply "helps" characterize the behavior of solutions of DifferentialEquations.jl is very undervaluing, in my eyes at least. This is perfectly clear from the contents list, https://juliadynamics.github.io/DynamicalSystems.jl/latest/#contents

@ahwillia Have you read the documentation of the submitted package, or at least the list of features offered by it?

Your initial question on the "differences" between the submitted library and DifferentialEquations.jl hinted that you have actually not read it, and the last comment only enhances this suspicion. I sincerely believe that this is not helpful in the current review process.
