### 2018-03-12 #598 @kyleniemeyer
@Datseris I would ask to tone down the language a little bit. It is a perfectly valid reviewer comment that the differences between this and your previous package are not entirely clear based on the article/README, and it should be fairly easy to improve the description to clear this up for a reader. The differences and justification are likely quite clear to you, but (evidently) not necessarily to someone else reading about the package.
