### 2018-03-13 #598 @ahwillia
Did not mean to undervalue the package at all. Something to "help" analyze dynamical systems is extremely useful. But it was unclear to me from reading the [article proof](https://github.com/openjournals/joss-papers/blob/joss.00598/joss.00598/10.21105.joss.00598.pdf) whether the goal of the library was analysis or simulation or a mix of both. I also looked at the README and first front page of documentation.

I think it is reasonable to point new users to `DifferentialEquations.jl` and explain when & why to use that library versus `DynamicalSystems.jl`. I don't think a new user should have to dive into the docstrings of specific methods/types to figure this out.

I've read through the documentation more deeply now and I've actually gone back and skimmed the "E&F Chaos" paper and related packages to try to get a better sense about what this package does and the motivations behind it. It's a very technical library and I won't pretend to deeply understand some of the features (though I actually do have some experience with nonlinear dynamical systems, perhaps more than the average newcomer). The documentation seems well referenced and I was able to download the library, pass unit tests, and reproduce examples in the documentation.

I expect this package documentation will be difficult for users without strong knowledge of this field. That being said, for those with the requisite background I could see this package being quite useful. It would be nice to incorporate more tutorials and pedagogical flow to the docs, but I do not expect this suggestion to be well received (and I do not view it as necessary). I do really appreciate the theorems which are sprinkled throughout the docs (e.g. the "shadowing theorem") and again the docs are very well-referenced with primary sources.

Consider me signed off.
