### 2018-03-13 #598 @Datseris
> I think it is reasonable to point new users to DifferentialEquations.jl and explain when & why to use that library versus DynamicalSystems.jl. I don't think a new user should have to dive into the docstrings of specific methods/types to figure this out.

@ahwillia Sure, this is a valid suggestion, but I can't think of a more natural point to do this, than the point that I introduce the concept that actually uses DifferentialEquations.jl. Doing it any point earlier does not seem very fitting, because as I said in my previous comment, only 1/10th of the exported functionality actually uses DifferentialEquations.jl.

Can you suggest a place where your suggestion would fit better?

> I expect this package will be difficult for users without strong knowledge of this field. That being said, for those with the requisite background I could see this package being quite useful. It would be nice to incorporate more tutorials and pedagogical flow to the docs

Thanks for pointing this out. My topmost priority with this package is to be more like a "library", and to be helpful for new users to also be able to *learn* new methods, and understand new methods. To this end I have tried to set up the following:

1. Every function is accompanied with a short description of the underlying algorithm.
2. References are given to scientific articles that first introduce the method or had significant influence on the method.
3. The documentation sections have one or two sentences about where and why you would like to apply the methods.
4. In many cases, caveats and difficulties are pointed out. Most of the time knowledge of these stuff comes from hands-on experience on working with a methods in scientific projects.

Would you mind telling me first how I can improve on this and, if possible, which points were not very clear for you when you went through it?

---

@kyleniemeyer

> @Datseris I would ask to tone down the language a little bit. It is a perfectly valid reviewer comment that the differences between this and your previous package are not entirely clear based on the article/README, and it should be fairly easy to improve the description to clear this up for a reader.

The reviewer did not comment at my previous package. I really do not understand your comment. "My" previous package (which is actually part of a GitHub organization and thus not associated with my name directly) is about billiard systems. But the reviewer never inquired about this, they inquired instead about the difference between the current submission and a solver of differential equations. I can't help but be confused. ???
