### 2018-03-13 JuliaDynamics/DynamicalSystems.jl#65 @Datseris
Distinguish clearly the three types of data structurs in introduction
Somewhere in the introduction we should point out that we use 3 data structurs (and various derivatives) in our packages.

1. continuous systems, that use DiffEq
2. discrete systems.
3. numerical datasets
