### 2018-03-13 JuliaDynamics/DynamicalSystems.jl#65 @Datseris
Scratch that, this is already done in the introduction, I have a quite clear numbered list.

Instead I should separate more the discrete systems and point out that continuous use solvers of differential equations, but are not actually connected with DiffEq other than that.
