### 2018-03-13 #598 @Datseris
Following @ahwillia I've opened the following issue: https://github.com/JuliaDynamics/DynamicalSystems.jl/issues/65 which I'll solve at the near future.
