### 2018-03-13 #598 @kyleniemeyer
@Datseris ah, I misremembered the bit about your prior package, sorry about that.

Aside from the issue you just mentioned, I have a few fixes for your paper:

- Could you add additional location details to your affiliation, such as city and country?
- The first reference, Strogatz, is missing bibliographic details—it just has an author and title. Is this an article or book? Either way, it needs some additional information (year, DOI, etc.)
- In the first sentence, the semicolon should be a comma.
- In the third sentence, "This vast extend" -> "This vast extent" (I think)
- In the first paragraph in the "Examples" section, there is a paragraph with "## Lyapunov spectrum of a continuous ..."; I think this was meant to be part of the code snippet that follows, rather than in the paragraph.
- The citation for DynamicalBilliards doesn't look quite right; perhaps remove the `publisher` field from the Bibtex entry?

Please let me know when you've made all those fixes, and also any remaining changes to the software. At that point, you will need to archive the repo (using Zenodo, for example) and provide the DOI here; the JOSS paper will be associated with whatever version that corresponds to.
