### 2018-03-13 #598 @ahwillia
I appreciate the desire to make the documentation more pedagogical. I really think it is a laudable goal and a great service to the field.

> Would you mind telling me first how I can improve on this and, if possible, which points were not very clear for you when you went through it?

It is great that each method includes a sentence or two description and again I really appreciate the amount of primary references and citations you included.

What I do not full understand (and this is partly my ignorance), is how a user should go about using all the functionality of this package at once on a single problem. That is, I would love to see a unifying example or vignette that pulls together the simulation, data analysis, and visualization tools. Let's say I have a new dynamical system that describes population dynamics of some species. What can your package tell me about it? What should I do first, and what should my follow-up questions be?

Perhaps putting a short roadmap/description about each of the six headings on this page would also help: https://juliadynamics.github.io/DynamicalSystems.jl/latest/chaos/overview/

To summarize, the documentation of individual functions is generally good, but I had a bit of trouble *connecting* concepts and figuring out how everything should synergistically work together.

Just my two cents. Thanks for all the effort you are putting into the package.
