### 2018-03-13 #598 @Datseris
@ahwillia

> What I do not full understand (and this is partly my ignorance), is how a user should go about using all the functionality of this package at once on a single problem. That is, I would love to see a unifying example or vignette that pulls together the simulation, data analysis, and visualization tools. Let's say I have a new dynamical system that describes population dynamics of some species. What can your package tell me about it? What should I do first, and what should my follow-up questions be?

Unfortunately it is not possible for me to answer this question(s). It would a great risk of misguiding someone into results they don't understand.

Each function of DynamicalSystems.jl has a well-defined purpose, mostly independent from the purpose of other functions (excluding of course cases where functions call each other). And each one can give you some information for your system. The question is, what information do you need? The user, as a scientist, should be the judge - I cannot help them in that decision.

I can only help by stating clearly what each method does (which at least in my eyes, I am fulfilling this goal, and also have an example application for any possible case). For further knowledge, one has to read papers and study. Having attended an introductory course in nonlinear dynamics and chaos is kind of mandatory. Just as having attended an introductory course in quantum mechanics is kind of mandatory for using a package about quantum mechanics, the same thing applies in using a package for nonlinear dynamics and chaos. It is an entire field of science after all (which also expands beyond physics and goes into chemistry, biology, economics, population dynamics, electronics and many more).

Mayhaps it is a good idea to show/cite some introductory textbooks on the field? I do that in the paper but not in the documentation.

(In the introduction, I do point out that the "target audience" of the library are scientists working on the field of nonlinear dynamics and chaos, in the goals section)

> but I had a bit of trouble connecting concepts and figuring out how everything should synergistically work together.

You had trouble connecting concepts exactly because there isn't any direct connection. They are a lot of tools for aiding in the study of dynamical systems, but they are not inter-connected in a "simple" manner. In fact, if you had the impression that everything works synergistically together, then I would have have written a bad documentation :D

P.S.: Many of the methods are actually connected in a rigorous and formal mathematical manner. I am stating this in case some expert ever reads this review. The connection is very deep though, sometimes with quite advanced math, and way beyond the scope of a documentation to discuss about.
