### 2018-03-14 #598 @Datseris
@kyleniemeyer I have done the changes you requested and tagged a new release.

The Zenodo DOI is: 10.5281/zenodo.1198278
