### 2018-03-14 #598 @kyleniemeyer
@Datseris most of the changes to the paper look good, though I noticed that the sentence "This vast extend of chaotic systems ..." in the first paragraph didn't get fixed; I believe "extend" should be "extent".
